# lab/testTask/documentation/api/openapi.yaml
openapi: 3.0.3
info:
  title: Shift Lab CRM API
  description: Описание CRM-системы для управления продавцами и их транзакциями.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Локальный сервер разработки
tags:
  - name: Sellers
    description: Управление продавцами
  - name: Transactions
    description: Управление транзакциями
  - name: Analytics
    description: Аналитические эндпоинты

paths:
  /api/sellers:
    get:
      tags: [Sellers]
      summary: Получить список всех продавцов
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SellerResponse'
    post:
      tags: [Sellers]
      summary: Создать нового продавца
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SellerRequest'
      responses:
        '201':
          description: Продавец успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SellerResponse'

  /api/sellers/{id}:
    get:
      tags: [Sellers]
      summary: Получить продавца по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SellerResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Sellers]
      summary: Обновить данные продавца
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SellerRequest'
      responses:
        '200':
          description: Данные успешно обновлены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SellerResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Sellers]
      summary: Удалить продавца
      description: Помечает продавца удаленным без физического уничтожения записей в БД.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Продавец успешно удален (No Content)
        '404':
          $ref: '#/components/responses/NotFound'

  /api/transactions:
    get:
      tags: [Transactions]
      summary: Получить список транзакций (с фильтром по продавцу)
      description: Возвращает все транзакции. Если передан параметр sellerId, возвращает транзакции только этого продавца.
      parameters:
        - name: sellerId
          in: query
          required: false
          description: ID продавца для фильтрации транзакций
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionResponse'
    post:
      tags: [Transactions]
      summary: Создать новую транзакцию
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionRequest'
      responses:
        '201':
          description: Транзакция успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'

  /api/transactions/{id}:
    get:
      tags: [Transactions]
      summary: Получить транзакцию по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/analytics/best-seller:
    get:
      tags: [Analytics]
      summary: Получить самого продуктивного продавца
      description: Вычисляет продавца с максимальной суммой транзакций за указанный период.
      parameters:
        - name: period
          in: query
          required: true
          description: Период для расчета (DAY, MONTH, QUARTER, YEAR)
          schema:
            type: string
            enum: [DAY, MONTH, QUARTER, YEAR]
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SellerResponse'

  /api/analytics/best-time/{sellerId}:
    get:
      tags: [Analytics]
      summary: Получить самое продуктивное время продавца
      parameters:
        - name: sellerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: object
                properties:
                  productivePeriod:
                    type: string
                    example: "14:00 - 15:00"

  /api/analytics/sellers/sum-less-than:
    get:
      tags: [Analytics]
      summary: Получить список продавцов с суммой меньше указанной
      description: Возвращает список продавцов, у которых сумма всех транзакций меньше указанного значения.
      parameters:
        - name: amount
          in: query
          required: true
          description: Пороговая сумма
          schema:
            type: number
            format: double
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SellerResponse'

components:
  schemas:
    SellerRequest:
      type: object
      required:
        - name
        - contactInfo
      properties:
        name:
          type: string
          example: "Иван Иванов"
        contactInfo:
          type: string
          example: "ivan@example.com, +79991234567"

    SellerResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Иван Иванов"
        contactInfo:
          type: string
          example: "ivan@example.com, +79991234567"
        registrationDate:
          type: string
          format: date-time
          example: "2023-10-25T10:00:00Z"

    TransactionRequest:
      type: object
      required:
        - sellerId
        - amount
        - paymentType
      properties:
        sellerId:
          type: integer
          format: int64
          example: 1
        amount:
          type: number
          format: double
          example: 1500.50
        paymentType:
          type: string
          enum: [CASH, CARD, TRANSFER]
          example: "CARD"

    TransactionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 100
        seller:
          $ref: '#/components/schemas/SellerResponse'
        amount:
          type: number
          format: double
          example: 1500.50
        paymentType:
          type: string
          example: "CARD"
        transactionDate:
          type: string
          format: date-time
          example: "2023-10-26T14:30:00Z"

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string

  responses:
    NotFound:
      description: Объект не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'